Add solutions for day 14
Technically my solution doesn't work out of the box. When initially
inserting characters, it will double count some characters. The reason
is because of how I calculate the number of letters. For example,
using the example polymer template, NNCB becomes NCNBCHB. I calculate
the number of C's as follows: count the number of pairs which contain a
C (1xNC, 1xCN, 1xBC, 1xCH), subtract from this the number of pairs in previous
polymers that result in a C (1xNN). The result is then 3 (4 - 1), but
should obviously only be 2. This discrepancy stays throughout the
execution (but never increases/decreases), so the actual answer can be
calculated by subtracting/adding as needed to the answer based on what
the max/min character counts are (in my case I needed to add 2 to my
answer because my H's were two too high).