Add solution for Day15
This is actually a pretty funny solution. I made an assumption that the
lowest risk for a cell must come from either above or from the left.
This isn't technically true since it's possible to have a chain of ones
or something that wrap around (for example) the cell to the left. I got
around this by also including looking to the bottom and to the right.
But instead of making a smart algorithm, it just reruns the calculations
multiple times. This algorithm technically fail for any test cases
designed to specifically counteract it. For example easiest worst case I
can think of is a chain of 17 1s that wrap around a chain of 9s like this:
1 9 9 9 x
1 9 9 9 1
1 9 9 1 1
1 9 9 1 9
1 9 9 1 9
1 9 9 1 9
1 9 9 1 9
1 1 1 1 9
Obviously that's assuming x is the final spot we want to get to. In this
case the path of 1s has a risk of 20, but my solution calculates it to
be 24. No matter how many times I make the algorithm rerun a test case
can still be constructed to beat it (just use the template above).
Luckily it worked though because their inputs are really simple 8)