Add solutions for Day 12
Probably not the cleanest solution, especially since it gets really slow
when the number of allowed repeated small cave visits gets above 3.
There's probably a more efficient way of doing this if you take into
consideration min cuts in the graph. That way movement isn't wasted on
going into a subgraph that can't be gotten out of.